+++ test/rust-test-2021/src/lib.rs
@@ -1,17 +1,22 @@
+#![cfg_attr(not(feature = "std"), no_std)]
 #![recursion_limit = "1024"]
+
+extern crate alloc;
 
 #[macro_use]
 pub mod macros;
 
 pub mod planus_api;
 pub mod planus_test;
+#[cfg(feature = "std")]
 pub mod planus_test_no_flatc;
 
+#[cfg(feature = "std")]
 pub mod hexdump;
 
 mod table_reader;
 
-#[cfg(test)]
+#[cfg(all(feature = "std", test))]
 pub mod tests {
     use std::path::Path;