+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Added extra unsafe blocks to templates to fix warnings for the 2024 edition
+- Updated tests for the 2024 edition
 
 ### Removed